Very WIP: Stage2 revival
Needs some corresponding changes in Cthulhu

WIP

WIP

WIP

WIP

fix inferred type and return handling of rrule calls (#83)

add custom Cthulhu toggle to show ad-transformed code (#85)

add inlining policy for ADInterpreter (#84)

fixup
